fix: build issues & small text api refactor
